requirements.txt
psycopg2-binary
Django
django-colorfield
Pillow
